LEP-1551
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@2692 05bdf26c-840f-0410-9ced-eb539d925f36

+++ util-java/src/com/liferay/util/BeanParamUtil.java
@@ -146,7 +146,7 @@ public class BeanParamUtil {
 			return ParamUtil.get(req, param, defaultValue);
 		}
 		else {
-			return ParamUtil.get(req, param, beanValue.intValue());
+			return ParamUtil.get(req, param, beanValue.longValue());
 		}
 	}
